ESP32 Core: Correct tool-related prefix and path

+++ arch/xtensa/include/types.h
@@ -69,8 +69,8 @@ typedef unsigned char      _uint8_t;
 typedef signed short       _int16_t;
 typedef unsigned short     _uint16_t;
 
-typedef signed int         _int32_t;
-typedef unsigned int       _uint32_t;
+typedef signed long        _int32_t;
+typedef unsigned long      _uint32_t;
 
 typedef signed long long   _int64_t;
 typedef unsigned long long _uint64_t;
@@ -78,12 +78,12 @@ typedef unsigned long long _uint64_t;
 
 /* A pointer is 4 bytes */
 
-typedef signed int         _intptr_t;
-typedef unsigned int       _uintptr_t;
+typedef signed long        _intptr_t;
+typedef unsigned long      _uintptr_t;
 
 /* This is the size of the interrupt state save returned by up_irq_save(). */
 
-typedef unsigned int       irqstate_t;
+typedef unsigned long      irqstate_t;
 
 #endif /* __ASSEMBLY__ */
